Convenient function for viewing any rectangular data file in an Excel window.

You can use this function for loading any data.frame, data_frame, tbl_df, matrix, vector objects into your system-default spreadsheet software (e.g., Excel). This function has been inspired by DataViewer and has implemented Hadley Wickham's write_csv instead of the default write.csv for a better performance.

view(x, ...)

Arguments

x
An object of class data.frame, matrix, or vector.

...
Any additional arguments available for write_csv.

Value

Data object opened in a preferrable spreadsheet application window.

Details

view

See example below.

Examples

library(ezviewr)
data(airquality)
str(airquality)

#> 'data.frame':	153 obs. of  6 variables:
#>  $ Ozone  : int  41 36 12 18 NA 28 23 19 8 NA ...
#>  $ Solar.R: int  190 118 149 313 NA NA 299 99 19 194 ...
#>  $ Wind   : num  7.4 8 12.6 11.5 14.3 14.9 8.6 13.8 20.1 8.6 ...
#>  $ Temp   : int  67 72 74 62 56 66 65 59 61 69 ...
#>  $ Month  : int  5 5 5 5 5 5 5 5 5 5 ...
#>  $ Day    : int  1 2 3 4 5 6 7 8 9 10 ...

## View your data object in your spreadsheet software:
view(airquality)

#> Warning: This function is only useful in interactive sessions running under Windows.

# Then, when necessary, you can modify the opened data in the spreadsheet and save it as a new data.